+++ simulator/node/CNS/Cns_enter.py
@@ -5,8 +5,8 @@ import math as m
 
 class Cns_enter(NodeC):
     def __init__(self,meta,_id, ligne_pwr,datalog):
-        self.max_power = int(meta['power'])
-        self.cost = int(meta['cost'])
+        self.max_power = float(meta['power'])
+        self.cost = float(meta['cost'])
         self.prior = 2
         self.randomArray = self.getRandomArray()
         super().__init__( _id, ligne_pwr,datalog)
