+++ simulator/node/CNS/Cns_town.py
@@ -5,8 +5,8 @@ import math as m
 
 class Cns_town(NodeC):
     def __init__(self,meta,_id, ligne_pwr,datalog):
-        self.max_power = int(meta['power'])
-        self.cost = int(meta['cost'])
+        self.max_power = float(meta['power'])
+        self.cost = float(meta['cost'])
         self.prior = 3
         self.randomArray = self.getRandomArray()
